+++ nodeconductor/logging/filters.py
@@ -12,8 +12,7 @@ from nodeconductor.core import serializers as core_serializers, filters as core_
 from nodeconductor.core.filters import ExternalFilterBackend
 from nodeconductor.logging import models, utils
 from nodeconductor.logging.elasticsearch_client import EmptyQueryset
-from nodeconductor.logging.loggers import event_logger
-from nodeconductor.logging.features import features_to_events, features_to_alerts, UPDATE_EVENTS
+from nodeconductor.logging.loggers import event_logger, expand_event_groups, expand_alert_groups
 
 
 def _convert(name):
@@ -50,12 +49,13 @@ class EventFilterBackend(filters.BaseFilterBackend):
         if 'event_type' in request.query_params:
             must_terms['event_type'] = request.query_params.getlist('event_type')
 
+        # Group events by features in order to prevent large HTTP GET request
         if 'exclude_features' in request.query_params:
             features = request.query_params.getlist('exclude_features')
-            must_not_terms['event_type'] = features_to_events(features)
+            must_not_terms['event_type'] = expand_event_groups(features)
 
         if 'exclude_extra' in request.query_params:
-            must_not_terms['event_type'] = must_not_terms.get('event_type', []) + UPDATE_EVENTS
+            must_not_terms['event_type'] = must_not_terms.get('event_type', []) + expand_event_groups(['update'])
 
         if 'user_username' in request.query_params:
             must_terms['user_username'] = [request.query_params.get('user_username')]
@@ -198,9 +198,10 @@ class AdditionalAlertFilterBackend(filters.BaseFilterBackend):
         if 'alert_type' in request.query_params:
             queryset = queryset.filter(alert_type__in=request.query_params.getlist('alert_type'))
 
+        # Group alerts by features in order to prevent large HTTP GET request
         if 'exclude_features' in request.query_params:
             features = request.query_params.getlist('exclude_features')
-            queryset = queryset.exclude(alert_type__in=features_to_alerts(features))
+            queryset = queryset.exclude(alert_type__in=expand_alert_groups(features))
 
         return queryset