+++ nodeconductor/structure/log.py
@@ -12,6 +12,10 @@ class CustomerEventLogger(EventLogger):
         event_types = ('customer_deletion_succeeded',
                        'customer_update_succeeded',
                        'customer_creation_succeeded')
+        event_groups = {
+            'customers': event_types,
+            'update': ('customer_update_succeeded',)
+        }
 
 
 class BalanceEventLogger(EventLogger):
@@ -21,6 +25,7 @@ class BalanceEventLogger(EventLogger):
     class Meta:
         event_types = ('customer_account_credited',
                        'customer_account_debited')
+        event_groups = {'customers': event_types}
 
 
 class ProjectEventLogger(EventLogger):
@@ -34,6 +39,13 @@ class ProjectEventLogger(EventLogger):
                        'project_update_succeeded',
                        'project_creation_succeeded',
                        'project_name_update_succeeded')
+        event_groups = {
+            'projects': event_types,
+            'update': (
+                'project_name_update_succeeded',
+                'project_update_succeeded'
+            )
+        }
 
 
 class ProjectGroupEventLogger(EventLogger):
@@ -43,6 +55,10 @@ class ProjectGroupEventLogger(EventLogger):
         event_types = ('project_group_deletion_succeeded',
                        'project_group_update_succeeded',
                        'project_group_creation_succeeded')
+        event_groups = {
+            'project_groups': event_types,
+            'update': ('project_group_update_succeeded',)
+        }
 
 
 class CustomerRoleEventLogger(EventLogger):
@@ -53,6 +69,7 @@ class CustomerRoleEventLogger(EventLogger):
 
     class Meta:
         event_types = 'role_granted', 'role_revoked'
+        event_groups = {'users': event_types}
 
 
 class ProjectRoleEventLogger(EventLogger):
@@ -65,6 +82,7 @@ class ProjectRoleEventLogger(EventLogger):
     class Meta:
         nullable_fields = ['project_group']
         event_types = 'role_granted', 'role_revoked'
+        event_groups = {'users': event_types}
 
 
 class ProjectGroupRoleEventLogger(EventLogger):
@@ -75,6 +93,7 @@ class ProjectGroupRoleEventLogger(EventLogger):
 
     class Meta:
         event_types = 'role_granted', 'role_revoked'
+        event_groups = {'users': event_types}
 
 
 class ProjectGroupMembershipEventLogger(EventLogger):
@@ -83,6 +102,7 @@ class ProjectGroupMembershipEventLogger(EventLogger):
 
     class Meta:
         event_types = 'project_added_to_project_group', 'project_removed_from_project_group'
+        event_groups = {'project_groups': event_types}
 
 
 class LicensesEventLogger(EventLogger):
@@ -103,6 +123,7 @@ class UserOrganizationEventLogger(EventLogger):
                        'user_organization_approved',
                        'user_organization_rejected',
                        'user_organization_removed')
+        event_groups = {'customers': event_types}
 
 
 class ResourceEventLogger(EventLogger):
@@ -133,6 +154,10 @@ class ResourceEventLogger(EventLogger):
             'resource_deletion_succeeded',
             'resource_deletion_failed',
         )
+        event_groups = {
+            'vms': event_types,
+            'update': ('resource_update_succeeded',)
+        }
 
 
 event_logger.register('customer_role', CustomerRoleEventLogger)